% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfit_ssm.r
\name{mfit_ssm}
\alias{mfit_ssm}
\title{mfit_ssm}
\usage{
mfit_ssm(d, ...)
}
\arguments{
\item{d}{input data.frame}

\item{...}{arguments to be passed to prefilter and fit_ssm}
}
\description{
Fit ssm to multiple individual tracks
}
