% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ctrw.r
\name{plot.ctrw}
\alias{plot.ctrw}
\title{plot}
\usage{
\method{plot}{ctrw}(m, page = 1)
}
\arguments{
\item{m}{a fitted object of class ctrw}

\item{page}{1 = plot all panels on a single page, 0 otherwise}
}
\description{
Visualise ctrw fit to track data
}
