% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ssm.R
\name{fit_ssm}
\alias{fit_ssm}
\title{Correlated Random Walk Filter}
\usage{
fit_ssm(d, tstep = 1, subset = rep(TRUE, nrow(d)), model = c("KF", "LS"),
  parameters = NULL, optim = c("nlminb", "optim"), verbose = FALSE,
  span = 0.1)
}
\arguments{
\item{d}{a data frame of observations including Argos KF error ellipse info}

\item{tstep}{the time step, in hours, to predict to}

\item{subset}{a logical vector indicating the subset of data records to be filtered}

\item{parameters}{the TMB parameter list}

\item{optim}{numerical optimizer}

\item{verbose}{report progress during minimization}

\item{span}{the span parameter for the loess fits used to estimate
initial locations}
}
\value{
a list with components
\item{\code{predicted}}{a data.frame of predicted location states}
\item{\code{fitted}}{a data.frame of fitted locations}
\item{\code{par}}{model parameter summmary}
\item{\code{data}}{the input data.frame}
\item{\code{subset}}{the input subset vector}
\item{\code{tstep}}{the prediction time step}
\item{\code{opt}}{the object returned by the optimizer}
\item{\code{tmb}}{the TMB object}
\item{\code{aic}}{the calculated Akaike Information Criterion}
}
\description{
Continuous-time Random Walk Filter
}
\details{
Fit a simple random walk in continuous time to filter Argos KF of LS data and predict
locations on a regular time step.

The input track is given as a dataframe where each row is an
observed location, with columns
\describe{
\item{'id'}{unique identifier,}
\item{'date'}{observation time (as GMT POSIXct),}
\item{'lc'}{location class,}
\item{'lon'}{observed longitude,}
\item{'lat'}{observed latitude}
}

The TMB parameter list can be specified directly with the
\code{parameters} argument. Otherwise suitable parameter values
are estimated from predicted locations estimated by fitting loess
smooths to the raw observations.

The filtering model assumes the errors in longitude and latitude
are proportional to scale factors determined by the location
class. The scale factors are specified through the \code{amf}
argument. By default the function uses the same scaling factors
for location accuracy as used in the R package crawl for ARGOS data.
}
