% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ssm.R
\name{fit_ssm}
\alias{fit_ssm}
\title{Correlated Random Walk Filter}
\usage{
fit_ssm(d, ts = 1, fit.to.subset = TRUE, parameters = NULL,
  optim = c("nlminb", "optim"), verbose = FALSE, span = 0.1)
}
\arguments{
\item{d}{a data frame of observations including Argos KF error ellipse info}

\item{ts}{the time step, in hours, to predict to}

\item{fit.to.subset}{a logical vector (default is TRUE) indicating whether the SSM is to be fit to the data subset determined by prefilter}

\item{optim}{numerical optimizer}

\item{verbose}{report progress during minimization}

\item{span}{the span parameter for the loess fits used to estimate
initial locations}
}
\value{
a list with components
\item{\code{predicted}}{a data.frame of predicted location states}
\item{\code{fitted}}{a data.frame of fitted locations}
\item{\code{par}}{model parameter summmary}
\item{\code{data}}{the input data.frame}
\item{\code{subset}}{the inpu subset vector}
\item{\code{ts}}{the prediction time step}
\item{\code{opt}}{the object returned by the optimizer}
\item{\code{tmb}}{the TMB object}
\item{\code{aic}}{the calculated Akaike Information Criterion}
}
\description{
Continuous-time Random Walk Filter
}
\details{
Fit a simple random walk in continuous time to filter Argos KF of LS data and predict
locations on a regular time step.
}
\examples{
\dontrun{
require(dplyr)
data(ellie)
## fit KF measurement model
fit <- ellie \%>\%
    select(1:8) \%>\%
    prefilter(., min.dist = 100) \%>\%
    fit_ssm(ts = 6)

## fit LS measurement model
fit.ls <- ellie \%>\%
    select(1:5) \%>\%
    prefilter(., min.dist = 100) \%>\%
    fit_ssm(ts = 6)
}

}
