% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefilter.r
\name{prefilter}
\alias{prefilter}
\title{Prepare Argos data for fitting a ctrw model}
\usage{
prefilter(d, span = 0.01, min.dist = 100, time.gap = NULL)
}
\arguments{
\item{d}{input data - must have 5 (LS), or 8 (KF) columns (see details)}

\item{span}{degree of loess smoothing (range: 0 - 1) to identify potential outliers}

\item{min.dist}{minimum distance from track to define potential outlier locations}

\item{time.gap}{not currently implemented}
}
\description{
\code{prefilter()} determines Argos data type (LS or KF); converts dates to POSIXt;
identifies observations with duplicate dates; orders observations in time; converts
longitudes from 0,360 to -180,180; projects lonlat coords to mercator x,y coords (in km);
adds location error multiplication factors based on Argos location class (for type LS);
and uses a loess smooth to identify potential outlier locations to be ignored when fitting
the \code{ctrw} model
}
\details{
Internal function
}
