% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfilter.R
\name{sfilter}
\alias{sfilter}
\title{fit ctrw SSM to individual track data}
\usage{
sfilter(d, ptime = 1, psi = 0, fit.to.subset = TRUE,
  parameters = NULL, optim = c("nlminb", "optim"), verbose = FALSE,
  f = 0.1, inner.control = NULL)
}
\description{
Filter Argos LS or KF track data
}
\details{
Internal function
}
